refactor(render): share SVG rendering with library consumers
