+++ crates/sdocx-cli/Cargo.toml
@@ -8,7 +8,6 @@ license = "GPL-3.0"
 repository = "https://github.com/twangodev/sdocx"
 
 [dependencies]
-sdocx = { version = "0.5.0", path = "../sdocx" }
+sdocx = { version = "0.5.0", path = "../sdocx", features = ["render"] }
 clap = { version = "4.5.60", features = ["derive"] }
-base64 = "0.22"
 resvg = "0.47.0"
