+++ crates/sdocx-wasm/Cargo.toml
@@ -11,6 +11,8 @@ repository = "https://github.com/twangodev/sdocx"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-sdocx = { version = "0.5.0", path = "../sdocx", features = ["serde"] }
+sdocx = { version = "0.5.0", path = "../sdocx", features = ["render", "serde"] }
 wasm-bindgen = "0.2"
+serde = { version = "1.0", features = ["derive"] }
 serde-wasm-bindgen = "0.6"
+sha2 = "0.10"
